backend/requirements.txt
Flask==3.1.2
flask-cors==5.0.0
python-dotenv==1.1.1
oracledb==2.5.0
bcrypt==4.2.1
PyJWT==2.10.1
pandas==2.2.0
openpyxl==3.1.2
gunicorn==21.2.0
